move code out of ark.py
